pyproject.toml
[project]
name = "peer-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
]

[project.scripts]
peer-agents = "peer_agents:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.40"]
google = ["google-generativeai"]
files = ["pypdf>=4.0", "python-docx>=1.0", "python-pptx>=1.0"]
all = ["openai>=1.0", "anthropic>=0.40", "google-generativeai", "pypdf>=4.0", "python-docx>=1.0", "python-pptx>=1.0"]